refactore check command
