+++ internal/domain/checkdom/handler.go
@@ -1,14 +1,24 @@
-package checkhandlers
+package checkdom
 
 import (
 	"encoding/json"
 	"fmt"
 	"net/http"
 
-	appdata "github.com/fiwon123/crower/internal/data/app"
+	"github.com/fiwon123/crower/internal/app"
 )
 
-func CheckNewVersion(currentVersion string, app *appdata.Data) (string, error) {
+type Handler struct {
+	app *app.Config
+}
+
+func NewHandler(app *app.Config) *Handler {
+	return &Handler{
+		app: app,
+	}
+}
+
+func (h *Handler) CheckNewVersion(currentVersion string) (string, error) {
 	resp, err := http.Get("https://api.github.com/repos/fiwon123/crower/releases/latest")
 	if err != nil {
 		return "", fmt.Errorf("error: %v \n", err)
